core: regenerate entities for the new Administrator table
