--- library/CoreBundle/Resources/config/lifecycle/ivoz/company_admin.yml
@@ -1,23 +1,0 @@
-services:
-  _defaults:
-    autowire: true
-    public: false
-
-  ##################################
-  ## Service collection
-  ##################################
-  Ivoz\Provider\Domain\Service\CompanyAdmin\CompanyAdminLifecycleServiceCollection:
-    public: true
-
-  provider.lifecycle.company_admin.service_collection:
-    '@Ivoz\Provider\Domain\Service\CompanyAdmin\CompanyAdminLifecycleServiceCollection'
-
-  ##################################
-  ## pre_persist
-  ##################################
-  Ivoz\Provider\Domain\Service\CompanyAdmin\CheckValidity:
-    tags:
-      - { name: provider.lifecycle.company_admin.pre_persist, priority: 10 }
-    arguments:
-      $companyAdminRepository:
-        '@=service("doctrine.orm.entity_manager").getRepository("Ivoz\\Provider\\Domain\\Model\\CompanyAdmin\\CompanyAdmin")'
