+++ library/Ivoz/Provider/Domain/Model/Administrator/AdministratorRepository.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Ivoz\Provider\Domain\Model\Administrator;
+
+use Doctrine\Common\Persistence\ObjectRepository;
+use Doctrine\Common\Collections\Selectable;
+
+interface AdministratorRepository extends ObjectRepository, Selectable {}
+
