--- library/Ivoz/Provider/Domain/Model/CompanyAdmin/CompanyAdminTrait.php
@@ -1,92 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Domain\Model\CompanyAdmin;
-
-use Ivoz\Core\Application\DataTransferObjectInterface;
-
-/**
- * CompanyAdminTrait
- * @codeCoverageIgnore
- */
-trait CompanyAdminTrait
-{
-    /**
-     * @var integer
-     */
-    protected $id;
-
-
-    /**
-     * Constructor
-     */
-    public function __construct()
-    {
-        parent::__construct(...func_get_args());
-
-    }
-
-    /**
-     * @return CompanyAdminDTO
-     */
-    public static function createDTO()
-    {
-        return new CompanyAdminDTO();
-    }
-
-    /**
-     * Factory method
-     * @param DataTransferObjectInterface $dto
-     * @return self
-     */
-    public static function fromDTO(DataTransferObjectInterface $dto)
-    {
-        /**
-         * @var $dto CompanyAdminDTO
-         */
-        $self = parent::fromDTO($dto);
-
-        if ($dto->getId()) {
-            $self->id = $dto->getId();
-            $self->initChangelog();
-        }
-
-        return $self;
-    }
-
-    /**
-     * @param DataTransferObjectInterface $dto
-     * @return self
-     */
-    public function updateFromDTO(DataTransferObjectInterface $dto)
-    {
-        /**
-         * @var $dto CompanyAdminDTO
-         */
-        parent::updateFromDTO($dto);
-
-        return $this;
-    }
-
-    /**
-     * @return CompanyAdminDTO
-     */
-    public function toDTO()
-    {
-        $dto = parent::toDTO();
-        return $dto
-            ->setId($this->getId());
-    }
-
-    /**
-     * @return array
-     */
-    protected function __toArray()
-    {
-        return parent::__toArray() + [
-            'id' => self::getId()
-        ];
-    }
-
-
-}
-
