--- library/Ivoz/Provider/Domain/Model/MainOperator/MainOperatorRepository.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Domain\Model\MainOperator;
-
-use Doctrine\Common\Persistence\ObjectRepository;
-use Doctrine\Common\Collections\Selectable;
-
-interface MainOperatorRepository extends ObjectRepository, Selectable {}
-
