--- library/Ivoz/Provider/Domain/Model/MainOperator/MainOperatorTrait.php
@@ -1,92 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Domain\Model\MainOperator;
-
-use Ivoz\Core\Application\DataTransferObjectInterface;
-
-/**
- * MainOperatorTrait
- * @codeCoverageIgnore
- */
-trait MainOperatorTrait
-{
-    /**
-     * @var integer
-     */
-    protected $id;
-
-
-    /**
-     * Constructor
-     */
-    public function __construct()
-    {
-        parent::__construct(...func_get_args());
-
-    }
-
-    /**
-     * @return MainOperatorDTO
-     */
-    public static function createDTO()
-    {
-        return new MainOperatorDTO();
-    }
-
-    /**
-     * Factory method
-     * @param DataTransferObjectInterface $dto
-     * @return self
-     */
-    public static function fromDTO(DataTransferObjectInterface $dto)
-    {
-        /**
-         * @var $dto MainOperatorDTO
-         */
-        $self = parent::fromDTO($dto);
-
-        if ($dto->getId()) {
-            $self->id = $dto->getId();
-            $self->initChangelog();
-        }
-
-        return $self;
-    }
-
-    /**
-     * @param DataTransferObjectInterface $dto
-     * @return self
-     */
-    public function updateFromDTO(DataTransferObjectInterface $dto)
-    {
-        /**
-         * @var $dto MainOperatorDTO
-         */
-        parent::updateFromDTO($dto);
-
-        return $this;
-    }
-
-    /**
-     * @return MainOperatorDTO
-     */
-    public function toDTO()
-    {
-        $dto = parent::toDTO();
-        return $dto
-            ->setId($this->getId());
-    }
-
-    /**
-     * @return array
-     */
-    protected function __toArray()
-    {
-        return parent::__toArray() + [
-            'id' => self::getId()
-        ];
-    }
-
-
-}
-
