--- library/Ivoz/Provider/Domain/Service/CompanyAdmin/CompanyAdminLifecycleEventHandlerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Domain\Service\CompanyAdmin;
-
-use Ivoz\Provider\Domain\Model\CompanyAdmin\CompanyAdminInterface;
-
-interface CompanyAdminLifecycleEventHandlerInterface
-{
-    public function execute(CompanyAdminInterface $entity, $isNew);
-}
