--- library/Ivoz/Provider/Domain/Service/CompanyAdmin/CompanyAdminLifecycleServiceCollection.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Domain\Service\CompanyAdmin;
-
-use Ivoz\Core\Domain\Service\LifecycleServiceCollectionInterface;
-use Ivoz\Core\Domain\Service\LifecycleServiceCollectionTrait;
-
-class CompanyAdminLifecycleServiceCollection implements LifecycleServiceCollectionInterface
-{
-    use LifecycleServiceCollectionTrait;
-
-    protected function addService(CompanyAdminLifecycleEventHandlerInterface $service)
-    {
-        $this->services[] = $service;
-    }
-}
