+++ library/Ivoz/Provider/Infrastructure/Persistence/Doctrine/AdministratorDoctrineRepository.php
@@ -3,14 +3,15 @@
 namespace Ivoz\Provider\Infrastructure\Persistence\Doctrine;
 
 use Doctrine\ORM\EntityRepository;
-use Ivoz\Provider\Domain\Model\MainOperator\MainOperatorRepository;
+use Ivoz\Provider\Domain\Model\Administrator\AdministratorRepository;
+
 
 /**
- * MainOperatorDoctrineRepository
+ * AdministratorDoctrineRepository
  *
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class MainOperatorDoctrineRepository extends EntityRepository implements MainOperatorRepository
+class AdministratorDoctrineRepository extends EntityRepository implements AdministratorRepository
 {
 }
