--- library/Ivoz/Provider/Infrastructure/Persistence/Doctrine/BrandOperatorDoctrineRepository.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Infrastructure\Persistence\Doctrine;
-
-use Doctrine\ORM\EntityRepository;
-use Ivoz\Provider\Domain\Model\BrandOperator\BrandOperatorRepository;
-
-/**
- * BrandOperatorDoctrineRepository
- *
- * This class was generated by the Doctrine ORM. Add your own custom
- * repository methods below.
- */
-class BrandOperatorDoctrineRepository extends EntityRepository implements BrandOperatorRepository
-{
-}
