--- library/Ivoz/Provider/Infrastructure/Persistence/Doctrine/CompanyAdminDoctrineRepository.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Ivoz\Provider\Infrastructure\Persistence\Doctrine;
-
-use Doctrine\ORM\EntityRepository;
-use Ivoz\Provider\Domain\Model\CompanyAdmin\CompanyAdminRepository;
-
-/**
- * CompanyAdminDoctrineRepository
- *
- * This class was generated by the Doctrine ORM. Add your own custom
- * repository methods below.
- */
-class CompanyAdminDoctrineRepository extends EntityRepository implements CompanyAdminRepository
-{
-}
