--- library/Ivoz/Provider/Infrastructure/Persistence/Doctrine/Mapping/MainOperator.MainOperator.orm.yml
@@ -1,16 +1,0 @@
-Ivoz\Provider\Domain\Model\MainOperator\MainOperator:
-  repositoryClass: Ivoz\Provider\Infrastructure\Persistence\Doctrine\MainOperatorDoctrineRepository
-  type: entity
-  table: MainOperators
-  id:
-    id:
-      type: integer
-      nullable: false
-      options:
-        unsigned: true
-      id: true
-      generator:
-        strategy: IDENTITY
-  lifecycleCallbacks:
-    postLoad:
-      - initChangelog
