--- library/Ivoz/Provider/Infrastructure/Persistence/Doctrine/Mapping/MainOperator.MainOperatorAbstract.orm.yml
@@ -1,60 +1,0 @@
-Ivoz\Provider\Domain\Model\MainOperator\MainOperatorAbstract:
-  type: mappedSuperclass
-  indexes:
-    timezoneId:
-      columns:
-        - timezoneId
-  uniqueConstraints:
-    username:
-      columns:
-        - username
-  fields:
-    username:
-      type: string
-      nullable: false
-      length: 65
-      options:
-        fixed: false
-    pass:
-      type: string
-      nullable: false
-      length: 80
-      options:
-        fixed: false
-        comment: '[password]'
-    email:
-      type: string
-      nullable: false
-      length: 100
-      options:
-        fixed: false
-        default: ''
-    active:
-      type: boolean
-      nullable: false
-      options:
-        default: '1'
-    name:
-      type: string
-      nullable: true
-      length: 100
-      options:
-        fixed: false
-    lastname:
-      type: string
-      nullable: true
-      length: 100
-      options:
-        fixed: false
-  manyToOne:
-    timezone:
-      targetEntity: \Ivoz\Provider\Domain\Model\Timezone\TimezoneInterface
-      cascade: {  }
-      fetch: LAZY
-      mappedBy: null
-      inversedBy: null
-      joinColumns:
-        timezoneId:
-          referencedColumnName: id
-          onDelete: set null
-      orphanRemoval: false
